+++ src/setIndex.js
@@ -1,5 +1,6 @@
 // @flow strict
 
+import {getAdjustedIndex} from './at.js';
 import {node} from './create.js';
 /*::
 import invariant from './invariant.js';
@@ -7,36 +8,74 @@ import type {
   ImmutableTree,
   NonEmptyImmutableTree,
 } from './types.js';
+
+type SetIndexOptions<T> =
+  | {
+      +index: number,
+      +value: T,
+      +updater: null,
+    }
+  | {
+      +index: number,
+      +value: null,
+      +updater: (existingTreeValue: T) => T,
+    };
 */
 
-function _setIndex/*:: <T> */(
+function _updateLeft/*:: <T> */(
   tree/*: NonEmptyImmutableTree<T> */,
-  index/*: number */,
-  value/*: T */,
   thisIndex/*: number */,
+  options/*: SetIndexOptions<T> */,
 )/*: NonEmptyImmutableTree<T> */ {
   const left = tree.left;
+  /*:: invariant(left.size !== 0); */
+  const newLeft = _setIndexInternal(left, thisIndex - left.right.size - 1, options);
+  if (newLeft === left) {
+    return tree;
+  }
+  return node(newLeft, tree.value, tree.right);
+}
+
+function _updateRight/*:: <T> */(
+  tree/*: NonEmptyImmutableTree<T> */,
+  thisIndex/*: number */,
+  options/*: SetIndexOptions<T> */,
+)/*: NonEmptyImmutableTree<T> */ {
   const right = tree.right;
-  const order = index - thisIndex;
+  /*:: invariant(right.size !== 0); */
+  const newRight = _setIndexInternal(right, thisIndex + right.left.size + 1, options);
+  if (newRight === right) {
+    return tree;
+  }
+  return node(tree.left, tree.value, newRight);
+}
+
+function _updateValue/*:: <T> */(
+  tree/*: NonEmptyImmutableTree<T> */,
+  options/*: SetIndexOptions<T> */,
+)/*: NonEmptyImmutableTree<T> */ {
+  const valueToInsert/*: T */ = options.updater === null
+    // $FlowIssue[incompatible-type]
+    ? options.value
+    : options.updater(tree.value);
+  if (Object.is(tree.value, valueToInsert)) {
+    return tree;
+  }
+  return node(tree.left, valueToInsert, tree.right);
+}
+
+export function _setIndexInternal/*:: <T> */(
+  tree/*: NonEmptyImmutableTree<T> */,
+  thisIndex/*: number */,
+  options/*: SetIndexOptions<T> */,
+)/*: NonEmptyImmutableTree<T> */ {
+  const order = options.index - thisIndex;
   if (order === 0) {
-    if (Object.is(tree.value, value)) {
-      return tree;
-    }
-    return node(left, value, right);
+    return _updateValue(tree, options);
   } else if (order < 0) {
-    /*:: invariant(left.size !== 0); */
-    const newLeft = _setIndex(left, index, value, thisIndex - left.right.size - 1);
-    if (newLeft === left) {
-      return tree;
-    }
-    return node(newLeft, tree.value, right);
+    return _updateLeft(tree, thisIndex, options);
   } else {
-    /*:: invariant(right.size !== 0); */
-    const newRight = _setIndex(right, index, value, thisIndex + right.left.size + 1);
-    if (newRight === right) {
-      return tree;
-    }
-    return node(left, tree.value, newRight);
+    return _updateRight(tree, thisIndex, options);
   }
 }
 
@@ -45,13 +84,14 @@ export default function setIndex/*:: <T> */(
   index/*: number */,
   value/*: T */,
 )/*: ImmutableTree<T> */ {
-  if (!Number.isInteger(index)) {
-    return tree;
-  }
-  let adjustedIndex = index < 0 ? (tree.size + index) : index;
-  if (adjustedIndex < 0 || adjustedIndex >= tree.size) {
+  let adjustedIndex = getAdjustedIndex(tree, index);
+  if (adjustedIndex === -1) {
     return tree;
   }
   /*:: invariant(tree.size !== 0); */
-  return _setIndex(tree, adjustedIndex, value, tree.left.size);
+  return _setIndexInternal(tree, tree.left.size, {
+    index: adjustedIndex,
+    value,
+    updater: null,
+  });
 }